+++ src/components/form-choice/index.test.js
@@ -8,6 +8,7 @@ describe("Choice element module", () => {
   it("renders empty string when undefined", () => {
     expect(Choice.makeComponent()).toBe("");
   });
+
   it("renders html-string of choice field", () => {
     expect(
       Choice.makeComponent({
@@ -17,9 +18,10 @@ describe("Choice element module", () => {
         type: "radio",
         checked: true,
         value: "z1",
-      })
+      }),
     ).toContain("Radio one label");
   });
+
   it("renders html-string of choice field with classing", () => {
     expect(
       Choice.makeComponent({
@@ -29,9 +31,10 @@ describe("Choice element module", () => {
         value: "z1",
         className: "test-btn btn-submit",
         classList: ["test-btn", "btn-submit"],
-      })
+      }),
     ).toContain("choice test-btn btn-submit");
   });
+
   it("renders html-string of class", () => {
     expect(
       Choice.makeComponent({
@@ -42,9 +45,10 @@ describe("Choice element module", () => {
         value: "z1",
         className: "test-btn btn-submit",
         classList: ["test-btn", "btn-submit"],
-      })
+      }),
     ).toContain("choice test-btn btn-submit");
   });
+
   it("removes # from id string and replaces it", () => {
     expect(
       Choice.makeComponent({
@@ -55,9 +59,10 @@ describe("Choice element module", () => {
         value: "z1",
         className: "test-btn btn-submit",
         classList: ["test-btn", "btn-submit"],
-      })
+      }),
     ).toContain('for="Z1"');
   });
+
   it("does not render when missing label", () => {
     expect(
       Choice.makeComponent({
@@ -67,9 +72,10 @@ describe("Choice element module", () => {
         value: "z1",
         className: "test-btn btn-submit",
         classList: ["test-btn", "btn-submit"],
-      })
+      }),
     ).toBe("");
   });
+
   it("does not render when missing id", () => {
     expect(
       Choice.makeComponent({
@@ -77,9 +83,10 @@ describe("Choice element module", () => {
         type: "option",
         value: "z1",
         selected: true,
-      })
+      }),
     ).toBe("");
   });
+
   it("returns an option html-string when type is option", () => {
     expect(
       Choice.makeComponent({
@@ -88,9 +95,10 @@ describe("Choice element module", () => {
         id: "Z1",
         type: "option",
         value: "z1",
-      })
+      }),
     ).toContain('<option value="z1" class="field"');
   });
+
   it("returns an option html-string when type is an option", () => {
     document.body.insertAdjacentHTML(
       "beforeend",
@@ -109,7 +117,7 @@ describe("Choice element module", () => {
         },
       ]
         .map(Choice.makeComponent)
-        .join("")
+        .join(""),
     );
 
     const radioInput = document.getElementById("Z0");
@@ -119,7 +127,7 @@ describe("Choice element module", () => {
         bubbles: true,
         cancelable: true,
         relatedTarget: radioInput,
-      })
+      }),
     );
     expect(radioParent.dataset.checked).toBe("true");
 
@@ -130,7 +138,7 @@ describe("Choice element module", () => {
         bubbles: true,
         cancelable: true,
         relatedTarget: checkboxInput,
-      })
+      }),
     );
     expect(checkboxParent.dataset.checked).toBe("true");
 
@@ -140,7 +148,7 @@ describe("Choice element module", () => {
         bubbles: true,
         cancelable: true,
         relatedTarget: bodyFrame,
-      })
+      }),
     );
     expect(bodyFrame.dataset.checked).toBeUndefined();
   });