+++ src/index.js
@@ -2,7 +2,9 @@ import * as dotenv from "dotenv";
 
 import Button from "./components/form-button";
 
-// TODO: setup `until-found` hidden states
+// TODO:
+//    setup `until-found` hidden states
+//    use `templates` api for HTML Modules
 
 dotenv.config();