+++ src/services/helper/index.js
@@ -15,9 +15,7 @@ export default {
   } = {}) {
     return this.booleanProperty(
       set && set.length,
-      Array.isArray(set)
-        ? set.map((txt) => `${prefix}${txt}${suffix}`).join(delimiter)
-        : `${prefix}${set}${suffix}`
+      Array.isArray(set) ? set.map((txt) => `${prefix}${txt}${suffix}`).join(delimiter) : `${prefix}${set}${suffix}`,
     );
   },
 
@@ -37,7 +35,7 @@ export default {
         ...acc,
         [key]: value && typeof value === "object" ? this.deepFreeze() : value,
       }),
-      {}
+      {},
     );
 
     return Object.freeze(obj);