+++ src/services/helper/index.test.js
@@ -27,40 +27,26 @@ describe("Util method", () => {
     });
 
     it("renders items of array", () => {
-      expect(Util.buildText({ set: ["btn:outlined", "btn:rounded"] })).toBe(
-        "btn:outlined btn:rounded"
-      );
+      expect(Util.buildText({ set: ["btn:outlined", "btn:rounded"] })).toBe("btn:outlined btn:rounded");
     });
 
     it("renders items of array with prefix", () => {
-      expect(
-        Util.buildText({ set: ["outlined", "rounded"], prefix: "btn:" })
-      ).toBe("btn:outlined btn:rounded");
+      expect(Util.buildText({ set: ["outlined", "rounded"], prefix: "btn:" })).toBe("btn:outlined btn:rounded");
     });
 
     it("renders item of string with suffix", () => {
-      expect(Util.buildText({ set: "btn:round", suffix: "ed" })).toBe(
-        "btn:rounded"
-      );
+      expect(Util.buildText({ set: "btn:round", suffix: "ed" })).toBe("btn:rounded");
     });
   });
 
   describe("booleanProperty", () => {
     it("renders value when condition succeeds", () => {
-      expect(
-        Util.booleanProperty(
-          true,
-          `<div class="${["test", "element"].join("-")}">Test</div>`
-        )
-      ).toBe('<div class="test-element">Test</div>');
+      expect(Util.booleanProperty(true, `<div class="${["test", "element"].join("-")}">Test</div>`)).toBe(
+        '<div class="test-element">Test</div>',
+      );
     });
     it("renders empty string  when condition fails", () => {
-      expect(
-        Util.booleanProperty(
-          false,
-          `<div class="${["test", "element"].join("-")}">Test</div>`
-        )
-      ).toBe("");
+      expect(Util.booleanProperty(false, `<div class="${["test", "element"].join("-")}">Test</div>`)).toBe("");
     });
   });
 
@@ -71,16 +57,14 @@ describe("Util method", () => {
           class: "class-name",
           id: "some-id",
           role: "button",
-        })
+        }),
       ).toBe('class="class-name" id="some-id" role="button"');
     });
   });
 
   describe("ArrayToString", () => {
     it("creates a deduplicated string from an array", () => {
-      expect(
-        Util.arrayToString(["a", "a", "b", "c", "d", 1, 2, "3", "3", 2, "c"])
-      ).toBe("a b c d 1 2 3");
+      expect(Util.arrayToString(["a", "a", "b", "c", "d", 1, 2, "3", "3", 2, "c"])).toBe("a b c d 1 2 3");
     });
   });